+++ src/main/java/callprotector/spring/service/ChatLogService/ChatLogService.java
@@ -1,12 +1,10 @@
-package callprotector.spring.service.ChatbotService;
+package callprotector.spring.service.ChatLogService;
 
 import callprotector.spring.web.dto.response.ChatLogResponseDTO;
 
 import java.util.List;
 
 public interface ChatLogService {
-
-    public ChatLogResponseDTO.ChatLogResponse askAndSaveChatLog(Long sessionId, String question);
 
     public List<ChatLogResponseDTO.ChatLogResponse> getChatLogsBySession(Long sessionId);
