+++ src/main/java/callprotector/spring/service/ChatSessionService/ChatSessionService.java
@@ -1,10 +1,15 @@
 package callprotector.spring.service.ChatSessionService;
 
+import callprotector.spring.domain.ChatSession;
+import callprotector.spring.domain.User;
 import callprotector.spring.web.dto.response.ChatSessionResponseDTO;
 
 
 public interface ChatSessionService {
 
-    ChatSessionResponseDTO.ChatSessionResponse createSession(String email);
+    public ChatSession getSessionById(Long sessionId);
+
+    ChatSessionResponseDTO.ChatSessionResponse createSession(User user);
+
 
 }
