+++ src/main/java/callprotector/spring/service/ChatSessionService/ChatSessionServiceImpl.java
@@ -3,7 +3,6 @@ package callprotector.spring.service.ChatSessionService;
 import callprotector.spring.domain.ChatSession;
 import callprotector.spring.domain.User;
 import callprotector.spring.repository.ChatSessionRepository;
-import callprotector.spring.repository.UserRepository;
 import callprotector.spring.web.dto.response.ChatSessionResponseDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,13 +16,17 @@ import java.time.LocalDateTime;
 public class ChatSessionServiceImpl implements ChatSessionService{
 
     private final ChatSessionRepository chatSessionRepository;
-    private final UserRepository userRepository;
+
 
     @Override
-    public ChatSessionResponseDTO.ChatSessionResponse createSession(String email) {
-        log.info("JWT 인증된 사용자 이메일: {}", email);
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자"));
+    public ChatSession getSessionById(Long sessionId) {
+        return chatSessionRepository.findById(sessionId)
+                .orElseThrow(() -> new IllegalArgumentException("세션을 찾을 수 없습니다."));
+    }
+
+    @Override
+    public ChatSessionResponseDTO.ChatSessionResponse createSession(User user) {
+        log.info("JWT 인증된 사용자 이메일: {}", user.getId());
 
         ChatSession session = ChatSession.builder()
                 .user(user)
