+++ src/main/java/callprotector/spring/service/UserService/UserService.java
@@ -1,5 +1,6 @@
 package callprotector.spring.service.UserService;
 
+import callprotector.spring.domain.User;
 import callprotector.spring.web.dto.request.UserRequestDTO;
 import callprotector.spring.web.dto.response.UserResponseDTO;
 
@@ -11,5 +12,7 @@ public interface UserService {
     public void sendVerificationCode(String email);
 
     public void verifyCode(String email, String code);
+
+    public User getUserByEmail(String email);
 
 }
