+++ src/main/java/callprotector/spring/web/controller/ChatLogController.java
@@ -1,7 +1,7 @@
 package callprotector.spring.web.controller;
 
 import callprotector.spring.apiPayload.ApiResponse;
-import callprotector.spring.service.ChatbotService.ChatLogService;
+import callprotector.spring.service.ChatLogService.ChatLogService;
 import callprotector.spring.web.dto.request.ChatbotRequestDTO;
 import callprotector.spring.web.dto.response.ChatLogResponseDTO;
 import lombok.RequiredArgsConstructor;
@@ -18,11 +18,6 @@ public class ChatLogController {
 
     private final ChatLogService chatLogService;
 
-    @PostMapping("/ask")
-    public ApiResponse<ChatLogResponseDTO.ChatLogResponse> askAndSaveLog(@RequestBody ChatbotRequestDTO.ChatbotRequest request) {
-        log.info("✅ ChatLogController 진입: sessionId={}, question={}", request.getSessionId(), request.getQuestion());
-        return ApiResponse.onSuccess(chatLogService.askAndSaveChatLog(request.getSessionId(), request.getQuestion()));
-    }
 
     @GetMapping("/session/{sessionId}")
     public ApiResponse<List<ChatLogResponseDTO.ChatLogResponse>> getLogs(@PathVariable Long sessionId) {
